requirements.txt
appdirs==1.4.4
asgiref==3.2.10
Babel==2.8.0
black==20.8b1
cffi==1.14.3
click==7.1.2
cryptography==3.2.1
Django==3.1.2
django-filer==2.0.2
django-js-asset==1.2.2
django-money==1.1
django-mptt==0.11.0
django-password-reset==2.0
django-phonenumber-field==5.0.0
django-polymorphic==3.0.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.1
easy-thumbnails==2.7
mypy-extensions==0.4.3
pathspec==0.8.0
phonenumberslite==8.12.11
Pillow==8.0.1
psycopg2==2.8.6
py-moneyed==0.8.0
pycparser==2.20
pytz==2020.1
regex==2020.10.23
reportlab==3.5.54
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
